Move stubs to test directory
