+++ src/app.ts
@@ -8,7 +8,7 @@ import * as dotenv from 'dotenv';
 import { LibraryDriver } from './drivers/LibraryDriver';
 import { MongoConnector } from './shared/Mongo/MongoConnector';
 import { LearningObjectAdapter } from './LearningObjects/LearningObjectAdapter';
-import { Stubs } from './stubs';
+import { Stubs } from './tests/stubs';
 dotenv.config();
 // ----------------------------------------------------------------------------------
 // Initializations
