+++ src/tests/stubs.ts
@@ -5,8 +5,8 @@ import {
     Outcome,
     LearningOutcome,
     StandardOutcome,
-} from './shared/entity';
-import { UserToken } from './shared/types';
+} from '../shared/entity';
+import { UserToken } from '../shared/types';
 
 // tslint:disable-next-line:no-require-imports
 let SEED_DATA = require('../test_environment/data');
